feat: replace hand-written CLI with cli-contracts generated code
- Move cli-contracts from devDependencies to dependencies (runtime policy derivation)
- Add TypeScript generator config to cli-contracts.config.yaml
- Auto-generate program.ts, policy.ts, types.ts, schemas.ts from cli-contract.yaml
- Create src/cli/handlers.ts implementing CommandHandlers interface
- Simplify cli.ts to use generated createProgram(handlers, version)
- Remove 9 hand-written command files under src/cli/commands/
- Remove manual --introspect from globalOptions (auto-added by effects)
- Regenerate cli-reference.md and TypeScript from updated contract
- Bump version to 0.22.0
